reorganizes test code, small docs fixes
